Remove hashpipe source files
This is now a HASHPIPE plugin project only.  It requires that HASHPIPE
be installed as a separate standalone program.

This commit also reworks configure.ac and Makefile.am to use the
HASHPIPE and xGPU autoconf macros that those packages install to
simplify creation of configure scripts.  See the updated INSTALL file
for more details.