Validate FID in paper_net_thread
The recevied FID should never be out of range unless an F engine is
misconfigured or there are bit errors in the FID field of the packet.
We have experienced the latter (bit errors in the FID field) and found
that invalid FIDs can lead to calculated pointers that point beyond the
end of the buffer to which they are supposed to point leading to to a
segfault.  Rejecting packets with invalid FIDs will prevent these
crashes, but it does not prevent bit errors that change a valid FID into
a different-but-still-valid FID from going to the wrong place in the
buffer.