Validate XID in packet header
Make sure we get a valid XID in the packet header.  If we don't, reject
that packet.  Invalid XIDs can be caused by misconfigured F engines,
incorrect 10 GbE network cabling, or bit errors that affect the XID
field of the header.  In theory, none of those things should ever
happen, but in practice they sometimes do.