Make paper_gpu_thread more robust to mcnt badness
Previously, paper_gpu_thread would only dump to the output buffer if its
input block's mcnt exactly equalled the expected "last_mcount" of the
integration.  There was a check to see whether the input block's was
greater than the epected last mcount, but it would only log such events
and not dump nor would it advance the expected last mcount.  If, for
some reason (such as severe network reliability issues) the expected
last mcount was never seen, then every mcount thereafter will be greater
than the never-to-be-updated-again expected last mcount and so no more
dumps would ever be passed on to the output buffer.  This effectively
stops all data dumps from the X engine instance!

This changes paper_gpu_thread such that it will dump if the input
block's mcnt is greater than or equal to the expected last mcount of the
integration.  It will still log an error message if the input mcnt is
greater than the expected value, but it will "recover" by dumping and
advancing the last mcount value.  This will still lead to two
consecutive corrupted dumps (the current dump that is "missing" the last
mcount but containing one block after it and the next dump that will be
missing at least the one block that got put into the current dump), but
subsequent dumps should be OK (assuming the cause of the missed mcount
does not persist).  If things are so bad that we miss the last mcount of
an integration, then the data are most likely already corrputed anyway
so this is not likely to make things significantly worse.