Make paper_net_thread start in a "hold off" state
When paper_net_thread starts running, it sets the status buffer key
"NETHOLD" to 1.  It then checks the value of this status buffer key once
every second until it has a value of 0.  Once it reads a value of 0 for
NETHOLD, it deletes the NETHOLD key and then proceeds to open the socket
and run normally.  This prevents the net thread of one instance from
trying to read network data while other instances are being started on
the same host.  Starting multiple instances has been shown to cause
dropped packets in started hashpipe processes while the other hashpipe
processes are still starting up.  Using the NETHOLD flag allows the
Linux processes to be started without causing dropped packets (because
no packets are being received) and the confusion (i.e. deadlocks) that
can occur when packets are received too far out of order.

The network thread has a lack of robustness to "missing" packets.  This
lack of robustness is really a bug that should be fixed, but this commit
provides a quick fix that alleviates the primary manifestation of that
bug.