Add paper_crc_thread and paper_crc_init.sh
paper_crc_thread is a net thread that simply reads the packets from the
network and verifies the CRC at the end of the packet.  Status buffer
keys beginning with "CRC" are created to display the statistics.
Specifically, the CRCPKOK and CRCPKERR fields show the number of packets
received without CRC errors (CRCPKOK) and the number of packets received
with CRC errors (CRCPKERR).

Because Zlib's crc32 function is not fast enough to keep up with the
PAPER dataflow, some packets are dropped.