Integrate TIME_DEMUX parameter
For HERA, TIME_DEMUX=2 (even/odd blocks of data are split between two distinct
correlators)

Using this parameter, rather than messing with the definition of MCNT, means
that MCNTs in each correlator reflect the spectrum index they are processing.
I.e., the spectrum index of the "even" correlator will have blocks with different
MCNTs to the "odd" correlator.
This should be taken into account when setting the start integration of each
correlator.

A first-pass python script to trigger integration starts (a la paper_ctl.rb)
has been added. It doesn't (yet) have a flag to figure out if it should
be setting the starting MCNT to the even or odd correlator.