"Working" catcher pipeline
The gathers data from both the even-sample and off-sample
correlators, averages over a specified number of channels (currently 4)
and saves the even/odd sum/diffs to HDF5 files (one file for the diffs,
one for the sums). This code transposes the baseline/channel axes, as requested
by RTP.
HDF5 headers are copied from a python-generated template, which is assumed to be
static over the course of an observation. Dynamic header items are generated
in C on the fly.

The performance of this code is quite erratic. Sometimes the netthread throughput
is ~20Gb/s. Sometimes it is ~2Gb/s and packets get dropped. Still under investigation.