Compute integration time / julian time arrays
Integration time needs to be fed into the catcher
from the correlator via the INTSECS status keyword.

Currently the julian time array is the same for
each block of Nbls. The integration_time array
is the same for every entry.

These arrays are placeholders for a more interesting future
where baseline-dependent averaging means the int-times
and time centers can be a function of baseline