Fixes for even/odd binning
This commit fixes a variety of bugs in the handling of MCNTs
when the correlator is separated into even and odd time slices.

MCNT rounding has been changed so as not to mask off the bits
which identify the correlator as even/odd. A status buffer variable
"TIMEIDX" has also been added, which tells a pipeline if it is expected
to be opperating on even or odd time chunks. If the received MCNTs
don't match this runtime-supplied variable, errors are logged.