Move channel averaging into the X-boxes
Previously, the catcher received all 8k channels
(at ~5Gb/s) and was required to average them down by a factor
of 4, transpose the channel and baseline axes, and write to disk.

The transpose is very slow, and would sometimes slow to the point it
would cause the netthread to lock waiting for a buffer.

This commit moves averaging into the X-engine, prior to data transmission,
meaning the catcher has to do much less work. This change has been put
in a new `hera_gpu_output_thread` with the hope that ultimately
the data transpose will be moved upstream as well.

Various tweaks to parameters have been made to make everything play nicely,
and correctly calculate buffer sizes / offsets / etc.