Don't write ant_[1|2]_array datasets
These need extending with each new accumulation, so should
be written by the receiver.

Instead, write a new dataset -- Header/corr_bl_order, which
is the baseline order out of the X-engines. The catcher
can manipulate this order to form the ant_[1|2]_array datasets
(potentially including stuff for baseline-dependent averaging)