Multithread baseline dependent averaging ...
... With a single thread processing BDA and output, the data rate
is set by the 2s cadence of buffers from the GPU thread becoming
available. To take advantage of the 16s cadence that can actually be
achieved, BDA needs to be multithreaded. This commit buffers data until
all baselines have 16s worth of computations, and an output thread that
then sends this data over a larger time period.